refactor(tests): consolidate duplicate MockVulnerabilityRepository into shared test_doubles module
Closes #539

Two incompatible MockVulnerabilityRepository definitions existed in separate test
modules: an always-empty unit struct in generate_sbom/tests.rs and a configurable
struct in check_vulnerabilities.rs.

Create src/application/use_cases/test_doubles.rs with a single
MockVulnerabilityRepository that supports both empty and pre-loaded variants via
new() and struct literal construction. Register the module behind #[cfg(test)] in
mod.rs and update both consumers to import from the shared location.
