+++ src/application/use_cases/mod.rs
@@ -3,6 +3,9 @@ mod check_vulnerabilities;
 mod fetch_licenses;
 mod generate_sbom;
 
+#[cfg(test)]
+pub(crate) mod test_doubles;
+
 #[allow(unused_imports)]
 pub use check_vulnerabilities::CheckVulnerabilitiesUseCase;
 pub use fetch_licenses::FetchLicensesUseCase;
